requirements.txt
Django~=3.1
psycopg2-binary~=2.8.5
djangorestframework~=3.11.1
celery~=4.4.7
django-celery-results~=1.2.1
django-celery-beat~=2.0.0
requests~=2.24.0
beautifulsoup4~=4.9.1
colorlog~=4.2.1